github2taiga/requirements.txt
github3
python-taiga
docopt
